分离文件读取和问价保存模块

+++ src/main/java/com/sun/xiaotian/diskmonitor/repository/FileSizeRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FileSizeRepository extends JpaRepository<FileSize, Long>, JpaSpecificationExecutor<FileSize> {
+public interface FileSizeRepository extends JpaRepository<FileSize, Integer>, JpaSpecificationExecutor<FileSize> {
 
 }
